+++ modules/tinymce/src/core/main/ts/api/OptionTypes.ts
@@ -147,7 +147,7 @@ interface BaseEditorOptions {
   language?: string;
   language_load?: boolean;
   language_url?: string;
-  lineheight_formats?: string;
+  line_height_formats?: string;
   max_height?: number;
   max_width?: number;
   menu?: Record<string, { title: string; items: string }>;